move getRequestAnimationFrame to util
