+++ components/_util/getRequestAnimationFrame.tsx
@@ -1,0 +1,9 @@
+export default function getRequestAnimationFrame() {
+  if (window.requestAnimationFrame) {
+    return window.requestAnimationFrame;
+  }
+  const prefix = ['moz', 'ms', 'webkit'].filter(key => `${key}RequestAnimationFrame` in window)[0];
+  return prefix
+    ? window[`${prefix}RequestAnimationFrame`]
+    : callback => setTimeout(callback, 1000 / 60);
+}
